create autodoc directory and move automodapi in
